Initial simplification with variables

+++ a1-extend-tiny-evaluator/tiny/eval.h
@@ -3,6 +3,6 @@
 
 #include "tree.h"
 
-double evalEXP(EXP *e);
+int evalEXP(EXP *e);
 
 #endif /* !EVAL_H */
